+++ src/tests/unittests/algorithms/test_random_telegraph_signal.py
@@ -7,7 +7,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from qtt.algorithms.random_telegraph_signal import (FittingException, _create_integer_histogram, generate_RTS_signal,
-                                                    plot_two_level_threshold, tunnelrates_RTS, two_level_threshold)
+                                                    tunnelrates_RTS)
 
 
 class TestRandomTelegraphSignal(unittest.TestCase):
@@ -78,10 +78,3 @@ class TestRandomTelegraphSignal(unittest.TestCase):
 
         with self.assertRaises(Exception):
             _ = _create_integer_histogram([])
-
-    def test_two_level_threshold(data):
-
-        data = generate_RTS_signal(100000, std_gaussian_noise=0.24, rate_up=10e3, rate_down=20e3)
-        result = two_level_threshold(data)
-        plot_two_level_threshold(result, fig=1)
-        plt.close(1)
